add pre and post hook test
